% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FFT_DATASETS}
\alias{FFT_DATASETS}
\title{Names of the 14 FFTrees benchmark datasets}
\format{
An object of class \code{character} of length 14.
}
\usage{
FFT_DATASETS
}
\description{
These are the classification datasets bundled with the
\pkg{FFTrees} package (Phillips, Neth, Woike & Gaissmaier, 2017),
minus pre-split variants of `heartdisease` (`heart.train`, `heart.test`,
`heart.cost`).
}
\keyword{datasets}
