% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{m_acc}
\alias{m_acc}
\title{Accuracy}
\usage{
m_acc(y, pred)
}
\arguments{
\item{y}{Integer vector of binary truth labels (0/1).}

\item{pred}{Integer vector of binary predictions (0/1).}
}
\value{
Scalar accuracy.
}
\description{
Accuracy
}
