% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{m_f1}
\alias{m_f1}
\title{F1 score}
\usage{
m_f1(y, pred)
}
\arguments{
\item{y}{Integer vector of binary truth labels (0/1).}

\item{pred}{Integer vector of binary predictions (0/1).}
}
\value{
Scalar F1.
}
\description{
F1 score
}
